+++ tests/unit/parallel/segmented_algorithms/partitioned_vector_inclusive_scan2.cpp
@@ -344,6 +344,6 @@ void inclusive_scan_tests(std::vector<hpx::id_type>& localities)
 int main()
 {
     std::vector<hpx::id_type> localities = hpx::find_all_localities();
-    inclusive_scan_tests<long long>(localities);
+    inclusive_scan_tests<double>(localities);
     return hpx::util::report_errors();
 }
