+++ tests/unit/parallel/segmented_algorithms/partitioned_vector_transform_binary1.cpp
@@ -4,6 +4,10 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/hpx_main.hpp>
+#include <hpx/include/partitioned_vector_predef.hpp>
+#include <hpx/include/parallel_transform.hpp>
+#include <hpx/include/parallel_count.hpp>
+
 #include <hpx/util/lightweight_test.hpp>
 
 #include <vector>
@@ -15,7 +19,7 @@ int main()
 {
     std::vector<hpx::id_type> localities = hpx::find_all_localities();
 
-    transform_binary_tests<int, int, double>(localities);
+    transform_binary_tests<double, double, int>(localities);
 
     return hpx::util::report_errors();
 }
